Move from row-wise filter to set-wise filter
This change primarily moves from filtering results row-by-row to
filtering them by the whole set. This allows the library being used for
fuzzy searching to sort the results by the closeness of the match.
Speaking of, this change also moves from using fuzzy for fuzzy searching
to using fuse; the result quality is much higher.